Move and rename HttpClient to HTTPClientHelper
HTTPClientHelper is now internal and abstract.
